+++ usertests/PixTestSetPh.cc
@@ -6,6 +6,7 @@
 
 #include "PixTestSetPh.hh"
 #include "log.h"
+#include "constants.h"
 
 using namespace std;
 using namespace pxar;
@@ -335,9 +336,9 @@ void PixTestSetPh::doTest()
   for( size_t roc = 0; roc < nRocs; ++roc ) {
 
     for( int i = 0; i < 256; ++i ) {
-      int ph0 = hmin[roc]->GetBinContent(i+1);
+      int ph0 = static_cast<int>(hmin[roc]->GetBinContent(i+1));
       if( ph0 <= 0 ) continue;
-      int ph4 = hmax[roc]->GetBinContent(i+1);
+      int ph4 = static_cast<int>(hmax[roc]->GetBinContent(i+1));
       if( ph4 <= 0 ) continue;
       if( roc == 0 ) {
 	LOG(logINFO) << "VOffsetRO " << setw(3) << i
@@ -443,7 +444,7 @@ void PixTestSetPh::doTest()
 
   for( size_t roc = 0; roc < nRocs; ++roc ) {
     for( int i = 0; i < 256; ++i ) {
-      int ph = href[roc]->GetBinContent(i+1);
+      int ph = static_cast<int>(href[roc]->GetBinContent(i+1));
       if( ph <= 0 ) continue;
       if( ph < 210 ) {
 	viref_adc_opt[roc] = i;
@@ -716,8 +717,8 @@ void PixTestSetPh::doTest()
       << " at " << setw(2) << max_col
       << "," << setw(2) << max_row;
 
-    int range = max_ph - min_ph;
-    int mid = (max_ph + min_ph)/2;
+    int range = static_cast<int>(max_ph - min_ph);
+    int mid = static_cast<int>(max_ph + min_ph)/2;
     int iter = 0;
 
     // iterate:
@@ -776,8 +777,8 @@ void PixTestSetPh::doTest()
 	  max_ph = vpix9[ipx].value;
       fApi->_dut->testPixel( max_col, max_row, false, roc );
 
-      range = max_ph - min_ph;
-      mid = (max_ph + min_ph)/2;
+      range = static_cast<int>(max_ph - min_ph);
+      mid = static_cast<int>(max_ph + min_ph)/2;
 
     } // while
